Add creating new test raylib window in a separate thread, from the main application. The idea is to get rid of the html ui and create ui in imgui style with raylib, and use as many windows necessary to do the editor windows

+++ src/globalContext.c
@@ -48,7 +48,7 @@ typedef struct globalContext
     float m_ambient;
     uint32_t m_ambientLoc;
 
-    int m_outlineSize;
+    uint16_t m_outlineSize;
     gizmo_t m_gizmoSelected;
     bool m_bbxChecked : 1;
 } globalContext_t;
